blog-automation/requirements.txt
anthropic==0.39.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.31.0
lxml==5.1.0
